Update - Spring Cleaning
Made some changes to underlying data structures to limit
redundancy.  The varref struct was completely removed, as it
was being used to hold a single unsigned int while pointing to
the rest of the relevant var data was accessed via pointer.
The unsigned int stored in varref was moved to the var struct.
A large number of functions across the program relied on varref,
so these functions have been modified to reflect the removal of
varref.